Separate implementation into different file
